+++ printing/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.3
 * Add file save and view to example application
+* Convert print screen example to Widgets
 
 # 2.0.2
 * Fix example application